feat: space and aliases in theme
