+++ src/components/header.js
@@ -1,6 +1,8 @@
 import { Link } from "gatsby"
 import PropTypes from "prop-types"
 import React from "react"
+import styled from "styled-components"
+import { space } from "styled-system"
 
 const Header = ({ siteTitle }) => (
   <header
